WebsiteFiles/back-end/requirement.txt
pyserial==3.5
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1 